259 events cmd

+++ events/models.py
@@ -5,9 +5,48 @@ from django.utils import timezone
 from persons.models import Person
 from features.models import Feature
 from django.core.validators import MinValueValidator
+from django.db.models import Q
+
+
+class OneTimeEventsManager(models.Manager):
+    def get_queryset(self):
+        return (
+            super()
+            .get_queryset()
+            .filter(parent=None)
+            .exclude(
+                pk__in=list(
+                    Event.objects.all().values_list("parent", flat=True).distinct()
+                )
+            )
+        )
+
+
+class ParentTrainingsManager(models.Manager):
+    def get_queryset(self):
+        return (
+            super()
+            .get_queryset()
+            .filter(
+                parent=None,
+                pk__in=list(
+                    Event.objects.all().values_list("parent", flat=True).distinct()
+                ),
+            )
+        )
+
+
+class ChildTrainingsManager(models.Manager):
+    def get_queryset(self):
+        return super().get_queryset().filter(~Q(parent=None))
 
 
 class Event(models.Model):
+    objects = models.Manager()
+    one_time_events = OneTimeEventsManager()
+    parent_trainings = ParentTrainingsManager()
+    child_trainings = ChildTrainingsManager()
+
     class State(models.TextChoices):
         FUTURE = "neuzavrena", _("neuzavřena")
         FINISHED = "uzavrena", _("uzavřena")
@@ -32,10 +71,9 @@ class Event(models.Model):
         "positions.EventPosition", through="events.EventPositionAssignment"
     )
     participants = models.ManyToManyField(Person, through="events.EventParticipation")
-    requirements = models.ManyToManyField(Feature, through="events.EventRequirement")
 
     def _is_top(self):
-        return self.parent == None
+        return self.parent is None
 
     def _is_top_training(self):
         children = Event.objects.filter(parent__exact=self)
@@ -106,7 +144,8 @@ class Event(models.Model):
         return self.name
 
     def delete(self, using=None, keep_parents=False):
-        self.price_list.delete()
+        if self.price_list is not None:
+            self.price_list.delete()
         return super().delete(using, keep_parents)
